#737 revert check_miovision and add new miovision_check
